pyproject.toml
[project]
name = "certificate-emission"
version = "0.1.0"
description = "Creates a PDF based on a docx file for CEUE activities."
authors = [
    {name = "Felipe Kauer",email = "fmkauer@gmail.com"}
]
license = {text = "MIT License"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "docx2pdf (>=0.1.8,<0.2.0)",
    "python-docx (>=1.1.2,<2.0.0)",
    "ruff (>=0.9.7,<0.10.0)",
    "polars (>=1.23.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
